Move initialize function for lunar gravity
